Add JWT and Cookies Support
We're using JWT since it satisfies our requriements for authentication and
cookies becuase we need a way of communicating JWTs that is simple(ish).
Other approaches won't work or are too complicated.

Add JWT to the application
    - Create a new Kubernetes secret and update the relevant config files
    to include the secret
    - Add functionality to sign JWT's with the secret

Add Cookie support to the application to store JWT's that are created
    - JWTs are stored in the `session` object attached to the request object
    - For the purposes of this application, we aren't encrypting the cookie

Normalize the output of MongoDB documents when fetching them as JSON.
    - Since there is no reason or guarantee that all the services in the
    application will use the same DB architecture, we need to make sure that
    the format of information is the same where ever possible