Create Components and Pages for Auth Routes
Due to the way this particular app is setup (all of the services will be
running in Docker containers in Kubernetes pods), the client (the Next.js
server) can't directly communicate with our other services. We need to tell
Axios the domain for the Ingress-Nginx instance when we make calls from
the Next.js server instead of the browser.
    - The domain name is the the namespace of Ingress-Nginx running in
    Kubernetes
    - The full address must have the form:
        'http://<NAMESPACE>.<SERVICE NAME>.svc.cluster.local'
    - Make sure to pass the request headers since the server side of the
    client does not do it for you like the browser

Create page components for signing in, signing up and signing out.

Create a header component that will be rendered for every page.
    - This component is placed once inside the 'root' app component wrapper
    that Next.js has

With the `getInitialProps` static method that components can have, when we
call it from the 'root' app component, we need to make sure that we also
call the `getInitialProps` methods that may exist for other components.