Create Skeleton for Tickets Service
The tickets service will be very similar in structure to the auth service.
The Dockerfile, Kubernetes configurations (including the ones for MongoDB),
and NPM dependencies were copied over and modified appropriately.

One change that was added to both the tickets and auth service was the use
of environment variables to describe the MongoDB URIs used to connect to
the databases.