Create Expiration Service
When an order is created, it has a 15 minute 'time to live'/'expiration
time'. Due to limitations with NATS Streaming, there is no way to attach
this time to live on messages, so a different approach needs to be
taken.

When an order is created, the expiration service listens for an
'order:created' event. When it recieves one, it uses Bull.js to send a
job to a Redis server which (after the expiration time) sends the job
back to Bull.js where it is processed. Once the job has been processed
an 'expiration:complete' event is published.

- Adds basic code structure for the expiration service
- Adds Kubernetes deployment files for both the expiration service and
the redis server that the expiration service will use
    - The expiration service is fairly bare-bones, so it does not
    require as much as the other services
- Updates skaffold setup